EnergyEvaluator/backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
pypdf
python-docx
google-generativeai
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
